requirements.txt
Django
django-environ
PyMySQL
djangorestframework